feat : JPS 최종 완성

+++ Assets/Path Finding/Scripts/NodeManager.cs
@@ -36,7 +36,7 @@ public class NodeManager : MonoBehaviour
 
     private void Start()
     {
-        nodes = new Node[(int)maxX, (int)maxY];
+        nodes = new Node[maxX, maxY];
         MakeTiles();
     }
